+++ examples/data-sources/netapp-ontap_storage_quota_rules/data-source.tf
@@ -1,0 +1,8 @@
+data "netapp-ontap_quota_rules" "storage_quota_rules" {
+  # required to know which system to interface with
+  cx_profile_name = "cluster2"
+  filter = {
+    type = "tree"
+    svm_name = "carchi-test"
+  }
+}